+++ src/app/core/@ngrx/products/products.state.ts
@@ -1,0 +1,15 @@
+import { IProductItem } from 'src/app/products/models/product.model';
+
+export interface ProductsState {
+    data: ReadonlyArray<IProductItem>;
+    readonly loading: boolean;
+    readonly loaded: boolean;
+    readonly error: Error | string;
+}
+
+export const initialProductsState: ProductsState = {
+    data: [],
+    loading: false,
+    loaded: false,
+    error: '',
+};