+++ src/app/core/@ngrx/router/router.actions.ts
@@ -1,0 +1,8 @@
+import { createAction, props } from '@ngrx/store';
+import { NavigationExtras } from '@angular/router';
+
+export const back = createAction('[Router] BACK');
+export const go = createAction(
+    '[Router] GO',
+    props<{ path: any[]; queryParams?: object; extras?: NavigationExtras }>(),
+);