+++ changelog.md
@@ -64,3 +64,11 @@
 ### Modified
 - `CartListComponent`, `ProductListComponent`, `ProductViewComponent`, `OrdersGuard`
 
+# Task7 - NgRx
+
+### Added
+- `@ngrx/store, @ngrx/effects` fot Products
+- `@ngrx/router-store`, implement navigation
+
+### Modified
+- `AppModule`, `CartListComponent`, `ProductListComponent`, `ProductViewComponent`, `ProductComponent`, `AdminGuard`, `ProductsPromiseService`